feat: add pie chart & move other charts
