+++ src/components/Dashboard.tsx
@@ -15,8 +15,8 @@ import { useEffect, useState } from "react";
 import YearDropdown from "@/components/YearDropdown";
 import { StatCard } from "@/components/ui/stat-card";
 import { ENTITY_CONFIG } from "@/lib/entity-config";
-import MultiLineGraph from "./LineGraph";
-import type { GraphDataset } from "./LineGraph";
+import MultiLineGraph from "./charts/LineGraph";
+import type { GraphDataset } from "./charts/LineGraph";
 import Link from "next/link";
 import { DashboardSkeleton } from "@/components/skeletons/DashboardSkeleton";
 import { LoadError } from "@/components/ui/load-error";
